refactor: コードレビュー指摘の修正 - build_summary をベースコントローラーに抽出 - ピボットロジックを CustomerTypePivot モジュールに共通化 - 未使用の @period_start/@period_end を削除 - DailyDetailsController のフルモンスクエリを直接クエリに変更 - Tailwind クラス名補間をリテラルに修正 - grand_total をメモ化
