+++ app/controllers/sales_analyses/base_controller.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module SalesAnalyses
+  class BaseController < ApplicationController
+    private
+
+    def build_summary
+      location = Location.find(params[:location_id])
+      period = (params[:period] || 30).to_i
+      Sales::AnalysisSummary.new(
+        location: location,
+        from: period.days.ago.beginning_of_day,
+        to: Time.current
+      )
+    end
+  end
+end
