+++ app/controllers/sales_analyses/cross_tables_controller.rb
@@ -1,24 +1,11 @@
 # frozen_string_literal: true
 
 module SalesAnalyses
-  class CrossTablesController < ApplicationController
+  class CrossTablesController < BaseController
     def show
-      summary = build_summary
       render SalesAnalyses::CrossTable::Component.new(
-        data: summary.cross_table
+        data: build_summary.cross_table
       ), layout: false
-    end
-
-    private
-
-    def build_summary
-      location = Location.find(params[:location_id])
-      period = (params[:period] || 30).to_i
-      Sales::AnalysisSummary.new(
-        location: location,
-        from: period.days.ago.beginning_of_day,
-        to: Time.current
-      )
     end
   end
 end
