+++ app/controllers/sales_analyses/summaries_controller.rb
@@ -1,24 +1,11 @@
 # frozen_string_literal: true
 
 module SalesAnalyses
-  class SummariesController < ApplicationController
+  class SummariesController < BaseController
     def show
-      summary = build_summary
       render SalesAnalyses::SummaryCards::Component.new(
-        data: summary.summary_by_customer_type
+        data: build_summary.summary_by_customer_type
       ), layout: false
-    end
-
-    private
-
-    def build_summary
-      location = Location.find(params[:location_id])
-      period = (params[:period] || 30).to_i
-      Sales::AnalysisSummary.new(
-        location: location,
-        from: period.days.ago.beginning_of_day,
-        to: Time.current
-      )
     end
   end
 end
