+++ app/views/components/sales_analyses/filter_bar/component.html.erb
@@ -19,7 +19,7 @@
     <%# 期間ボタン %>
     <div class="join">
       <% period_options.each do |p| %>
-        <%= link_to "#{p}日", period_path(p), class: "#{period_class(p)} join-item" %>
+        <%= link_to "#{p}日", period_path(p), class: period_class(p) %>
       <% end %>
     </div>
   </div>
